API completed

+++ Backend/YanKoltukBackend/YanKoltukBackend/Services/Implementations/ManagerService.cs
@@ -22,26 +22,28 @@ namespace YanKoltukBackend.Services.Implementations
             var managers = await _managerRepo.GetAllAsync();
             var manager = managers.FirstOrDefault(m => m.UserId == userId);
             if (manager == null)
-            {
-                return ServiceResult<int>.ErrorResult("Manager not found");
-            }
+                return ServiceResult<int>.ErrorResult("Error: Manager not found");
             return ServiceResult<int>.SuccessResult(manager.ManagerId);
         }
 
-        public async Task<List<Service>> GetAllServicesAsync()
+        public async Task<List<Service>> GetAllServicesAsync(int managerId)
         {
-            return (await _serviceRepo.GetAllAsync()).ToList();
+            return (await _serviceRepo.FindAsync(s => s.ManagerId == managerId)).ToList();
         }
 
-        public async Task<Service> GetServiceByIdAsync(int serviceId)
+        public async Task<Service?> GetServiceByIdAsync(int managerId, int serviceId)
         {
-            return await _serviceRepo.GetByIdAsync(serviceId);
+            return (await _serviceRepo.FindAsync(s => (s.ManagerId == managerId) && (s.ServiceId == serviceId))).FirstOrDefault();
         }
 
         public async Task<ServiceResult<Service>> AddServiceAsync(ServiceDto serviceDto, int managerId)
         {
             try
             {
+                var manager = await _managerRepo.GetByIdAsync(managerId);
+                if (manager == null)
+                    return ServiceResult<Service>.ErrorResult("Error: Manager not found");
+
                 var passwd = AuthHelper.GeneratePasswd();
                 var user = _userHelper.CreateUser(serviceDto.Plate, passwd, Roles.Service.GetDescription());
                 await _userRepo.AddAsync(user);
@@ -54,6 +56,7 @@ namespace YanKoltukBackend.Services.Implementations
                     DepartureLocation = serviceDto.DepartureLocation,
                     DepartureTime = serviceDto.DepartureTime,
                     ManagerId = managerId,
+                    Manager = manager,
                     DriverIdNo = serviceDto.DriverIdNo,
                     DriverName = serviceDto.DriverName,
                     DriverPhone = serviceDto.DriverPhone,
@@ -64,12 +67,6 @@ namespace YanKoltukBackend.Services.Implementations
                     StewardessPhoto = serviceDto.StewardessPhoto
                 };
                 await _serviceRepo.AddAsync(service);
-
-                var manager = await _managerRepo.GetByIdAsync(managerId);
-                if (manager == null)
-                {
-                    return ServiceResult<Service>.ErrorResult("Manager not found");
-                }
 
                 manager.Services.Add(service);
                 await _managerRepo.UpdateAsync(manager);
@@ -82,41 +79,60 @@ namespace YanKoltukBackend.Services.Implementations
             }
         }
 
-        public async Task<ServiceResult<Service>> UpdateServiceAsync(UpdateServiceDto updateServiceDto, Service service)
+        public async Task<ServiceResult<Service>> UpdateServiceAsync(UpdateServiceDto updateServiceDto, int managerId, int serviceId)
         {
             try
             {
-                service.DriverIdNo = updateServiceDto.DriverIdNo;
-                service.DriverName = updateServiceDto.DriverName;
-                service.DriverPhone = updateServiceDto.DriverPhone;
-                service.DriverPhoto = updateServiceDto.DriverPhoto;
+                var manager = await _managerRepo.GetByIdAsync(managerId);
+                if (manager == null)
+                    return ServiceResult<Service>.ErrorResult("Error: Manager not found");
 
-                service.StewardessIdNo = updateServiceDto.StewardessIdNo;
-                service.StewardessName = updateServiceDto.StewardessName;
-                service.StewardessPhone = updateServiceDto.StewardessPhone;
-                service.StewardessPhoto = updateServiceDto.StewardessPhoto;
+                var service = await GetServiceByIdAsync(managerId, serviceId);
+                if (service == null)
+                    return ServiceResult<Service>.ErrorResult("Error: Service not found");
+
+                if (!string.IsNullOrWhiteSpace(updateServiceDto.DriverIdNo))
+                    service.DriverIdNo = updateServiceDto.DriverIdNo;
+                if (!string.IsNullOrWhiteSpace(updateServiceDto.DriverName))
+                    service.DriverName = updateServiceDto.DriverName;
+                if (!string.IsNullOrWhiteSpace(updateServiceDto.DriverPhone))
+                    service.DriverPhone = updateServiceDto.DriverPhone;
+                if (!string.IsNullOrWhiteSpace(updateServiceDto.DriverPhoto))
+                    service.DriverPhoto = updateServiceDto.DriverPhoto;
+
+                if (!string.IsNullOrWhiteSpace(updateServiceDto.StewardessIdNo))
+                    service.StewardessIdNo = updateServiceDto.StewardessIdNo;
+                if (!string.IsNullOrWhiteSpace(updateServiceDto.StewardessName))
+                    service.StewardessName = updateServiceDto.StewardessName;
+                if (!string.IsNullOrWhiteSpace(updateServiceDto.StewardessPhone))
+                    service.StewardessPhone = updateServiceDto.StewardessPhone;
+                if (!string.IsNullOrWhiteSpace(updateServiceDto.StewardessPhoto))
+                    service.StewardessPhoto = updateServiceDto.StewardessPhoto;
 
                 await _serviceRepo.UpdateAsync(service);
                 return ServiceResult<Service>.SuccessResult(service, "Service updated");
             }
             catch (Exception ex)
             {
-                return ServiceResult<Service>.ErrorResult("Error updating service: " + ex.Message);
+                return ServiceResult<Service>.ErrorResult("Error: Service not updated - " + ex.Message);
             }
         }
 
-        public async Task<ServiceResult<Service>> DeleteServiceAsync(int serviceId)
+        public async Task<ServiceResult<Service>> DeleteServiceAsync(int managerId, int serviceId)
         {
-            var service = await _serviceRepo.GetByIdAsync(serviceId);
+            var manager = await _managerRepo.GetByIdAsync(managerId);
+            if (manager == null)
+                return ServiceResult<Service>.ErrorResult("Error: Manager not found");
+
+            var service = await GetServiceByIdAsync(managerId, serviceId);
+            if (service == null)
+                return ServiceResult<Service>.ErrorResult("Error: Service not found");
+
             var user = await _userRepo.GetByIdAsync(service.UserId);
 
-            if (service == null)
-            {
-                return ServiceResult<Service>.ErrorResult("Service not found");
-            }
-
             await _serviceRepo.DeleteAsync(service);
-            await _userRepo.DeleteAsync(user);
+            if (user != null)
+                await _userRepo.DeleteAsync(user);
 
             return ServiceResult<Service>.SuccessResult(service);
         }
